+++ src/Laravel/Reporters/LaravelLogReporter.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Henzeb\Enumhancer\Laravel\Reporters;
 
 use BackedEnum;
@@ -9,11 +10,11 @@ use function class_basename;
 class LaravelLogReporter implements Reporter
 {
 
-    public function report(string $enum, string $key, ?BackedEnum $context): void
+    public function report(string $enum, ?string $key, ?BackedEnum $context): void
     {
         Log::warning(
             class_basename($enum)
-            . ' does not have \'' . $key . '\'',
+            . ($key ? ' does not have \'' . $key . '\'' : ': A null value was passed'),
             array_filter([
                 'class' => class_basename($enum),
                 'key' => $key,