+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `Enumhancer` will be documented in this file
 
+## 1.0.1 - 2022 02-16
+
+- You can now define a mapper in a method
+- When you use an empty string or null in mappable, it will return null now.
+
 ## 1.0.0 - 2022-02-15
 
 - initial release